+++ server/.env
@@ -1,2 +1,3 @@
 DB_URL="mongodb+srv://vscodemissong:mortyparty@mancode.ku2ubmu.mongodb.net/wallsync-app?retryWrites=true&w=majority"
 PORT=8000
+JWT_SECRET = 123456789
