+++ server/config/db.js
@@ -2,11 +2,11 @@ const mongoose = require("mongoose");
 
 async function connectDB() {
 	try {
+		mongoose.set("strictQuery", false);
 		const connect = await mongoose.connect(process.env.DB_URL, {
 			useNewUrlParser: true,
 			useUnifiedTopology: true,
 		});
-		mongoose.set("strictQuery", false);
 		console.log(`Connected to WallSync DB: ${connect.connection.host}`);
 	} catch (err) {
 		console.log(err);
