+++ server/routes/contributionRoute.js
@@ -1,41 +1,10 @@
 const express = require("express");
-const Group = require("../models/billSplit/Group");
-const Contribution = require("../models/billSplit/Contribution");
-const User = require("../models/user/User");
+const contributionController = require("../controllers/contributionController");
 const router = express.Router();
 
-router.post("/:groupName/:firstName/addCont", (req, res) => {
-	Group.findOne({ groupName: req.params.groupName })
-		.then((gd) => {
-			User.findOne({ firstName: req.params.firstName })
-				.then((user) => {
-					const newCont = new Contribution(req.body);
-
-					newCont
-						.save()
-						.then((data) => {
-							gd.contributions.push({
-								name: data.contributedBy,
-								desc: data.description,
-								share: data.amount,
-							});
-							gd.groupTotal += data.amount;
-							gd.save();
-							res
-								.status(201)
-								.json({ contri: gd.contributions, total: gd.groupTotal });
-						})
-						.catch((err) => {
-							res.json("error1");
-						});
-				})
-				.catch((err) => {
-					res.json("error2");
-				});
-		})
-		.catch((err) => {
-			res.json("error3");
-		});
-});
+router.post(
+	"/:groupName/:firstName/addCont",
+	contributionController.addContribution
+);
 
 module.exports = router;
