+++ server/server.js
@@ -6,7 +6,8 @@ const mongoose = require("mongoose");
 const connectDB = require("./config/db");
 const cors = require("cors");
 const userRoute = require("./routes/userRoute");
-
+const groupRoute = require("./routes/groupRoute");
+const contributionRoute = require("./routes/contributionRoute");
 const app = express();
 const PORT = process.env.PORT || 8008;
 connectDB();
@@ -14,6 +15,9 @@ connectDB();
 app.use(cors());
 app.use(express.json());
 app.use("/api", userRoute);
+app.use("/api", groupRoute);
+app.use("/api", contributionRoute);
+
 app.listen(PORT, () => {
 	console.log("WallSync is running on port " + PORT);
 });
